fix build on OpenBSD/i386 (and maybe other 32-bits architectures ?)
Otherwise the build of 7.7.3 fails with:
utilities/transactions/pessimistic_transaction.cc:687:25: error: implicit conversion from
    'std::numeric_limits<unsigned long long>::type' (aka 'unsigned long long') to 'size_t' (aka 'unsigned long') changes value from 18446744073709551615 to 4294967295 [-Werror,-Wconstant-conversion]
       : std::numeric_limits<uint64_t>::max();
        ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
utilities/transactions/pessimistic_transaction.cc:762:23: error: implicit conversion from
    'std::numeric_limits<unsigned long long>::type' (aka 'unsigned long long') to 'size_t' (aka 'unsigned long') changes value from 18446744073709551615 to 4294967295 [-Werror,-Wconstant-conversion]
       : std::numeric_limits<uint64_t>::max();

+++ utilities/transactions/pessimistic_transaction.cc
@@ -684,7 +684,7 @@ Status WriteCommittedTxn::CommitWithoutPrepareInternal() {
           const Comparator* ucmp =
               WriteBatchWithIndexInternal::GetUserComparator(*wbwi, cf);
           return ucmp ? ucmp->timestamp_size()
-                      : std::numeric_limits<uint64_t>::max();
+                      : std::numeric_limits<size_t>::max();
         });
     if (!s.ok()) {
       return s;
@@ -759,7 +759,7 @@ Status WriteCommittedTxn::CommitInternal() {
         const Comparator* ucmp =
             WriteBatchWithIndexInternal::GetUserComparator(*wbwi, cf);
         return ucmp ? ucmp->timestamp_size()
-                    : std::numeric_limits<uint64_t>::max();
+                    : std::numeric_limits<size_t>::max();
       });
     }
   }
